requirements.txt
apturl==0.5.2
awscli==1.11.13
beautifulsoup4==4.4.1
blinker==1.3
botocore==1.4.70
Brlapi==0.6.4
cachetools==2.0.1
certifi==2017.4.17
chardet==3.0.4
checkbox-support==0.22
click==3.3
colorama==0.3.7
command-not-found==0.3
cryptography==1.2.3
decorator==4.1.2
defer==1.0.6
Django==1.11.2
django-debug-toolbar==1.8
djangorestframework==3.6.3
docutils==0.12
feedparser==5.1.3
firebase-admin==2.3.0
google-auth==1.0.2
google-cloud-core==0.27.1
google-cloud-storage==1.4.0
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.8
jmespath==0.9.0
language-selector==0.1
lml==0.0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
odfpy==1.3.5
olefile==0.44
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==4.1.1
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.1.1
pycups==1.9.73
pycurl==7.43.0
pyexcel==0.5.3
pyexcel-io==0.4.4
pyexcel-ods==0.4.1
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.5.0
pyparsing==2.0.3
PySocks==1.6.7
python-apt==1.1.0b1
python-dateutil==2.4.2
python-debian==0.1.27
python-firebase==1.2
python-systemd==231
pytz==2017.2
pyxdg==0.25
reportlab==3.3.0
requests==2.18.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
sqlparse==0.2.3
sticky==0.3
sublime==0.1.0
system-service==0.3
texttable==0.9.1
traitlets==4.3.2
twilio==6.3.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.21.1
usb-creator==0.3.0
virtualenv==15.1.0
wcwidth==0.1.7
xdiagnose==3.8.4.1
xkit==0.0.0
xlrd==1.1.0
XlsxWriter==0.7.3
